update from original

+++ src/Content/ContentManager.py
@@ -23,11 +23,11 @@ class ContentManager(object):
         self.contents = ContentDbDict(site)
         self.hashfield = PeerHashfield()
         self.has_optional_files = False
-        self.site.onFileDone.append(lambda inner_path: self.addOptionalFile(inner_path))
 
+    # Load all content.json files
     def loadContents(self):
         if len(self.contents) == 0:
-            self.log.debug("Content db not initialized, load files from filesystem")
+            self.log.debug("ContentDb not initialized, load files from filesystem")
             self.loadContent(add_bad_files=False, delete_removed_files=False)
         self.site.settings["size"] = self.getTotalSize()
 
@@ -35,10 +35,11 @@ class ContentManager(object):
         if "hashfield" in self.site.settings.get("cache", {}):
             self.hashfield.fromstring(self.site.settings["cache"]["hashfield"].decode("base64"))
             del self.site.settings["cache"]["hashfield"]
-            self.has_optional_files = True
         elif self.contents.get("content.json") and self.getOptionalSize() > 0:
             self.site.storage.updateBadFiles()  # No hashfield cache created yet
-            self.has_optional_files = True
+        self.has_optional_files = bool(self.hashfield)
+
+        self.contents.db.initSite(self.site)
 
     # Load content.json to self.content
     # Return: Changed files ["index.html", "data/messages.json"], Deleted files ["old.jpg"]
@@ -63,10 +64,10 @@ class ContentManager(object):
 
                 new_content = json.load(open(content_path))
             except Exception, err:
-                self.log.error("%s load error: %s" % (content_path, Debug.formatException(err)))
+                self.log.warning("%s load error: %s" % (content_path, Debug.formatException(err)))
                 return [], []
         else:
-            self.log.error("Content.json not exist: %s" % content_path)
+            self.log.warning("Content.json not exist: %s" % content_path)
             return [], []  # Content.json not exist
 
         try:
@@ -95,18 +96,18 @@ class ContentManager(object):
                 if old_content and old_content.get("files_optional", {}).get(relative_path):
                     # We have the file in the old content
                     old_hash = old_content["files_optional"][relative_path].get("sha512")
-                    if old_hash != new_hash and self.site.settings.get("autodownloadoptional"):
-                        changed.append(content_inner_dir + relative_path)  # Download new file
-                    elif old_hash != new_hash and not self.site.settings.get("own"):
+                    if old_hash != new_hash and self.site.isDownloadable(file_inner_path):
+                        changed.append(file_inner_path)  # Download new file
+                    elif old_hash != new_hash and self.hashfield.hasHash(old_hash) and not self.site.settings.get("own"):
                         try:
-                            self.hashfield.removeHash(old_hash)
+                            self.optionalRemove(file_inner_path, old_hash, old_content["files_optional"][relative_path]["size"])
                             self.site.storage.delete(file_inner_path)
                             self.log.debug("Deleted changed optional file: %s" % file_inner_path)
                         except Exception, err:
                             self.log.debug("Error deleting file %s: %s" % (file_inner_path, err))
                 else:  # The file is not in the old content
-                    if self.site.settings.get("autodownloadoptional"):
-                        changed.append(content_inner_dir + relative_path)  # Download new file
+                    if self.site.isDownloadable(file_inner_path):
+                        changed.append(file_inner_path)  # Download new file
 
             # Check deleted
             if old_content:
@@ -120,12 +121,20 @@ class ContentManager(object):
                     **new_content.get("files_optional", {})
                 )
 
-                deleted = [content_inner_dir + key for key in old_files if key not in new_files]
+                deleted = [key for key in old_files if key not in new_files]
                 if deleted and not self.site.settings.get("own"):
                     # Deleting files that no longer in content.json
-                    for file_inner_path in deleted:
+                    for file_relative_path in deleted:
+                        file_inner_path = content_inner_dir + file_relative_path
                         try:
                             self.site.storage.delete(file_inner_path)
+
+                            # Check if the deleted file is optional
+                            if old_content.get("files_optional") and old_content["files_optional"].get(file_relative_path):
+                                old_hash = old_content["files_optional"][file_relative_path].get("sha512")
+                                if self.hashfield.hasHash(old_hash):
+                                    self.optionalRemove(file_inner_path, old_hash, old_content["files_optional"][file_relative_path]["size"])
+
                             self.log.debug("Deleted file: %s" % file_inner_path)
                         except Exception, err:
                             self.log.debug("Error deleting file %s: %s" % (file_inner_path, err))
@@ -148,6 +157,7 @@ class ContentManager(object):
             if old_content and "user_contents" in new_content and "archived" in new_content["user_contents"]:
                 old_archived = old_content.get("user_contents", {}).get("archived", {})
                 new_archived = new_content.get("user_contents", {}).get("archived", {})
+                self.log.debug("old archived: %s, new archived: %s" % (len(old_archived), len(new_archived)))
                 archived_changed = {
                     key: date_archived
                     for key, date_archived in new_archived.iteritems()
@@ -159,6 +169,7 @@ class ContentManager(object):
                         archived_inner_path = content_inner_dir + archived_dirname + "/content.json"
                         if self.contents.get(archived_inner_path, {}).get("modified", 0) < date_archived:
                             self.removeContent(archived_inner_path)
+                    self.site.settings["size"] = self.getTotalSize()
 
             # Load includes
             if load_includes and "includes" in new_content:
@@ -195,10 +206,13 @@ class ContentManager(object):
             new_content["signs"] = None
             if "cert_sign" in new_content:
                 new_content["cert_sign"] = None
+
+            if new_content.get("files_optional"):
+                self.has_optional_files = True
             # Update the content
             self.contents[content_inner_path] = new_content
         except Exception, err:
-            self.log.error("Content.json parse error: %s" % Debug.formatException(err))
+            self.log.warning("Content.json parse error: %s" % Debug.formatException(err))
             return [], []  # Content.json parse error
 
         # Add changed files to bad files
@@ -245,25 +259,24 @@ class ContentManager(object):
         except Exception, err:
             self.log.debug("Error key from contents: %s" % inner_path)
 
-
     # Get total size of site
     # Return: 32819 (size of files in kb)
     def getTotalSize(self, ignore=None):
-        size = self.contents.db.getTotalSize(self.site.address, ignore)
+        size = self.contents.db.getTotalSize(self.site, ignore)
         if size:
             return size
         else:
             return 0
 
     def getOptionalSize(self):
-        size = self.contents.db.getOptionalSize(self.site.address)
+        size = self.contents.db.getOptionalSize(self.site)
         if size:
             return size
         else:
             return 0
 
     def listModified(self, since):
-        return self.contents.db.listModified(self.site.address, since)
+        return self.contents.db.listModified(self.site, since)
 
     def listContents(self, inner_path="content.json", user_files=False):
         back = [inner_path]
@@ -460,12 +473,15 @@ class ContentManager(object):
             if ignored:  # Ignore content.json, defined regexp and files starting with .
                 self.log.info("- [SKIPPED] %s" % file_relative_path)
             else:
-                file_path = self.site.storage.getPath(dir_inner_path + "/" + file_relative_path)
+                file_inner_path = dir_inner_path + "/" + file_relative_path
+                file_path = self.site.storage.getPath(file_inner_path)
                 sha512sum = CryptHash.sha512sum(file_path)  # Calculate sha512 sum of file
                 if optional:
                     self.log.info("- [OPTIONAL] %s (SHA512: %s)" % (file_relative_path, sha512sum))
-                    files_optional_node[file_relative_path] = {"sha512": sha512sum, "size": os.path.getsize(file_path)}
-                    self.hashfield.appendHash(sha512sum)
+                    file_size = os.path.getsize(file_path)
+                    files_optional_node[file_relative_path] = {"sha512": sha512sum, "size": file_size}
+                    if not self.hashfield.hasHash(sha512sum):
+                        self.optionalDownloaded(file_inner_path, sha512sum, file_size, own=True)
                 else:
                     self.log.info("- %s (SHA512: %s)" % (file_relative_path, sha512sum))
                     files_node[file_relative_path] = {"sha512": sha512sum, "size": os.path.getsize(file_path)}
@@ -622,7 +638,7 @@ class ContentManager(object):
         name, domain = content["cert_user_id"].split("@")
         cert_address = rules["cert_signers"].get(domain)
         if not cert_address:  # Cert signer not allowed
-            self.log.error("Invalid cert signer: %s" % domain)
+            self.log.warning("Invalid cert signer: %s" % domain)
             return False
         return CryptBitcoin.verify(
             "%s#%s/%s" % (rules["user_address"], content["cert_auth_type"], name), cert_address, content["cert_sign"]
@@ -636,52 +652,58 @@ class ContentManager(object):
         old_content = self.contents.get(inner_path)
         if old_content:
             old_content_size = len(json.dumps(old_content, indent=1)) + sum([file["size"] for file in old_content.get("files", {}).values()])
+            old_content_size_optional = sum([file["size"] for file in old_content.get("files_optional", {}).values()])
         else:
             old_content_size = 0
-
+            old_content_size_optional = 0
 
         content_size_optional = sum([file["size"] for file in content.get("files_optional", {}).values()])
         site_size = self.site.settings["size"] - old_content_size + content_size  # Site size without old content plus the new
+        site_size_optional = self.site.settings["size_optional"] - old_content_size_optional + content_size_optional  # Site size without old content plus the new
 
         site_size_limit = self.site.getSizeLimit() * 1024 * 1024
 
         # Check site address
         if content.get("address") and content["address"] != self.site.address:
-            self.log.error("%s: Wrong site address: %s != %s" % (inner_path, content["address"], self.site.address))
+            self.log.warning("%s: Wrong site address: %s != %s" % (inner_path, content["address"], self.site.address))
             return False
 
         # Check file inner path
         if content.get("inner_path") and content["inner_path"] != inner_path:
-            self.log.error("%s: Wrong inner_path: %s" % (inner_path, content["inner_path"]))
+            self.log.warning("%s: Wrong inner_path: %s" % (inner_path, content["inner_path"]))
             return False
 
         # Check total site size limit
         if site_size > site_size_limit:
-            self.log.error("%s: Site too large %s > %s, aborting task..." % (inner_path, site_size, site_size_limit))
+            self.log.warning("%s: Site too large %s > %s, aborting task..." % (inner_path, site_size, site_size_limit))
+            if inner_path == "content.json" and self.site.settings["size"] == 0:
+                # First content.json download, save site size to display warning
+                self.site.settings["size"] = site_size
             task = self.site.worker_manager.findTask(inner_path)
             if task:  # Dont try to download from other peers
                 self.site.worker_manager.failTask(task)
             return False
 
         if inner_path == "content.json":
-            self.site.settings["size"] = site_size  # Save to settings if larger
+            self.site.settings["size"] = site_size
+            self.site.settings["size_optional"] = site_size_optional
             return True  # Root content.json is passed
 
         # Load include details
         rules = self.getRules(inner_path, content)
         if not rules:
-            self.log.error("%s: No rules" % inner_path)
+            self.log.warning("%s: No rules" % inner_path)
             return False
 
         # Check include size limit
         if rules.get("max_size") is not None:  # Include size limit
             if content_size > rules["max_size"]:
-                self.log.error("%s: Include too large %s > %s" % (inner_path, content_size, rules["max_size"]))
+                self.log.warning("%s: Include too large %s > %s" % (inner_path, content_size, rules["max_size"]))
                 return False
 
         if rules.get("max_size_optional") is not None:  # Include optional files limit
             if content_size_optional > rules["max_size_optional"]:
-                self.log.error("%s: Include optional files too large %s > %s" % (
+                self.log.warning("%s: Include optional files too large %s > %s" % (
                     inner_path, content_size_optional, rules["max_size_optional"])
                 )
                 return False
@@ -690,21 +712,22 @@ class ContentManager(object):
         if rules.get("files_allowed"):
             for file_inner_path in content["files"].keys():
                 if not re.match("^%s$" % rules["files_allowed"], file_inner_path):
-                    self.log.error("%s %s: File not allowed" % (inner_path, file_inner_path))
+                    self.log.warning("%s %s: File not allowed" % (inner_path, file_inner_path))
                     return False
 
         if rules.get("files_allowed_optional"):
             for file_inner_path in content.get("files_optional", {}).keys():
                 if not re.match("^%s$" % rules["files_allowed_optional"], file_inner_path):
-                    self.log.error("%s %s: Optional file not allowed" % (inner_path, file_inner_path))
+                    self.log.warning("%s %s: Optional file not allowed" % (inner_path, file_inner_path))
                     return False
 
         # Check if content includes allowed
         if rules.get("includes_allowed") is False and content.get("includes"):
-            self.log.error("%s: Includes not allowed" % inner_path)
+            self.log.warning("%s: Includes not allowed" % inner_path)
             return False  # Includes not allowed
 
-        self.site.settings["size"] = site_size  # Save to settings if larger
+        self.site.settings["size"] = site_size
+        self.site.settings["size_optional"] = site_size_optional
 
         return True  # All good
 
@@ -719,7 +742,7 @@ class ContentManager(object):
                 else:
                     new_content = json.load(file)
                 if inner_path in self.contents:
-                    old_content = self.contents.get(inner_path)
+                    old_content = self.contents.get(inner_path, {"modified": 0})
                     # Checks if its newer the ours
                     if old_content["modified"] == new_content["modified"] and ignore_same:  # Ignore, have the same content.json
                         return None
@@ -731,10 +754,10 @@ class ContentManager(object):
                         # gevent.spawn(self.site.publish, inner_path=inner_path)  # Try to fix the broken peers
                         return False
                 if new_content["modified"] > time.time() + 60 * 60 * 24:  # Content modified in the far future (allow 1 day+)
-                    self.log.error("%s modify is in the future!" % inner_path)
+                    self.log.warning("%s modify is in the future!" % inner_path)
                     return False
                 if self.isArchived(inner_path, new_content["modified"]):
-                    self.log.error("%s this file is archived!" % inner_path)
+                    self.log.warning("%s this file is archived!" % inner_path)
                     return False
                 # Check sign
                 sign = new_content.get("sign")
@@ -756,11 +779,11 @@ class ContentManager(object):
                         if not CryptBitcoin.verify(
                             "%s:%s" % (signs_required, ",".join(valid_signers)), self.site.address, new_content["signers_sign"]
                         ):
-                            self.log.error("%s invalid signers_sign!" % inner_path)
+                            self.log.warning("%s invalid signers_sign!" % inner_path)
                             return False
 
                     if inner_path != "content.json" and not self.verifyCert(inner_path, new_content):  # Check if cert valid
-                        self.log.error("%s invalid cert!" % inner_path)
+                        self.log.warning("%s invalid cert!" % inner_path)
                         return False
 
                     valid_signs = 0
@@ -776,7 +799,7 @@ class ContentManager(object):
                     return CryptBitcoin.verify(sign_content, self.site.address, sign)
 
             except Exception, err:
-                self.log.error("Verify sign error: %s" % Debug.formatException(err))
+                self.log.warning("Verify sign error: %s" % Debug.formatException(err))
                 return False
 
         else:  # Check using sha512 hash
@@ -789,7 +812,7 @@ class ContentManager(object):
                 else:
                     hash_valid = False
                 if file_info.get("size", 0) != file.tell():
-                    self.log.error(
+                    self.log.warning(
                         "%s file size does not match %s <> %s, Hash: %s" %
                         (inner_path, file.tell(), file_info.get("size", 0), hash_valid)
                     )
@@ -797,14 +820,28 @@ class ContentManager(object):
                 return hash_valid
 
             else:  # File not in content.json
-                self.log.error("File not in content.json: %s" % inner_path)
+                self.log.warning("File not in content.json: %s" % inner_path)
                 return False
 
-    def addOptionalFile(self, inner_path):
-        info = self.getFileInfo(inner_path)
-        if info and info["optional"]:
-            self.log.debug("Downloaded optional file, adding to hashfield: %s" % inner_path)
-            self.hashfield.appendHash(info["sha512"])
+    def optionalDownloaded(self, inner_path, hash, size=None, own=False):
+        if size is None:
+            size = self.site.storage.getSize(inner_path)
+        if type(hash) is int:
+            done = self.hashfield.appendHashId(hash)
+        else:
+            done = self.hashfield.appendHash(hash)
+        self.site.settings["optional_downloaded"] += size
+        return done
+
+    def optionalRemove(self, inner_path, hash, size=None):
+        if size is None:
+            size = self.site.storage.getSize(inner_path)
+        if type(hash) is int:
+            done = self.hashfield.removeHashId(hash)
+        else:
+            done = self.hashfield.removeHash(hash)
+        self.site.settings["optional_downloaded"] -= size
+        return done
 
 
 if __name__ == "__main__":
